Separación datos en ropositories y creación de Factories en viewModels
