+++ app/src/main/java/com/example/myapplication4/ui/transaccion/TransactionViewModel.kt
@@ -3,6 +3,7 @@ package com.example.myapplication4.ui.transaccion
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import com.example.myapplication4.Clases.Transaccion
 import com.example.myapplication4.repository.TransactionRepository
@@ -57,4 +58,14 @@ class TransactionViewModel(private val repository: TransactionRepository) : View
 
 enum class Period {
     DAY, WEEK, MONTH, YEAR, ALL
+}
+
+class TransactionViewModelFactory(private val repository: TransactionRepository) : ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(TransactionViewModel::class.java)) {
+            @Suppress("UNCHECKED_CAST")
+            return TransactionViewModel(repository) as T
+        }
+        throw IllegalArgumentException("Unknown ViewModel class")
+    }
 }
